// eshop-libs/projects/sag-common/src/enums/affiliate.enum.ts
export enum AffiliateEnum {
    TECHNO_CH = 'technomag',
    DEREND_CH = 'derendinger-ch',
    DEREND_AT = 'derendinger-at',
    MATIK_AT = 'matik-at',
    MATIK_CH = 'matik-ch',
    WBB = 'wbb',
    RBE = 'rbe',
    EH_CH = 'eh-ch',
    EH_AT = 'eh-at',
    CZ = 'st-cz',
    SB = 'wt-sb',
    EH_CZ = 'eh-cz',
    AXCZ = 'ax-cz',
    EH_AX_CZ = 'eh-ax-cz'
}